WiP: download and save close to done
This version doesn't run, but the data structures and components for
downloading and saving are getting there, and the data structure to be
presented in the REST API is also not far away.

So what remains in the pipeline is the data pivot transformation. Since
I'm using SMHI's discoverability and there's a lot of URLs it doesn't
lend itself well to testing, and saving is one-shot.

The transformation itself should be a perfect candidate for testing
though. Will use TDD for that.